+++ types/quantum-circuit/tsconfig.json
@@ -13,6 +13,7 @@
         "forceConsistentCasingInFileNames": true
     },
     "files": [
-        "index.d.ts",
+        "circuit.d.ts",
+        "index.d.ts"
     ]
 }
